changed the datatype for credits in course options from BigDecimal to string as we might have ranges and multiples in credits

+++ contrib/myplan/trunk/ks-myplan/myplan-ui/src/test/java/org/kuali/student/ksap/schedulebuilder/ScheduleBuilderTest.java
@@ -251,7 +251,7 @@ public class ScheduleBuilderTest {
         courseOption.setCourseId(UUID.randomUUID().toString());
         courseOption.setCourseCode(campus + curric + courseNumber);
         courseOption.setCourseTitle(courseTitle);
-        courseOption.setCredits(BigDecimal.valueOf(Long.valueOf(credits)));  // always null in debugger?
+        courseOption.setCredits(credits);  // always null in debugger?
         courseOption.setDiscarded(false);
         courseOption.setLockedIn(false);
         courseOption.setSelected(false);
